Initial commit.
This is a utility program to simplify the process of building
and running various parts of Chrome.

This originated in
https://github.com/cmumford/cmumfordenv/tree/master/bin/crbuild
and was copied out at commit 768a885244e2ca348483358a36b6d9267e6e7728
into it's own repository as it is fairly large and unrelated to
the other tools in cmumfordenv.